==> 3_anal.R <==
rm(list = ls())
setwd("d://klips2017/data") 

#library(dummies)
library(sqldf)
library(lmtest)
library(multiwayvcov)

load("wage.Rdata")
load("bdata.Rdata")
pdata<-merge(wage, bdata, by="pid", all.x=T)
pdata$lhwage<-log(pdata$hwage)
pdata$lmwage<-log(pdata$mwage)
pdata$lhearnings<-log(pdata$hearnings)
pdata$female<-ifelse(pdata$sex==2, 1, 0)
pdata$area.f<-factor(pdata$area)
pdata$ind.f<-factor(pdata$ind)
pdata$ind2<-floor(pdata$ind/10)
pdata$ind2.f<-factor(pdata$ind2)
pdata$occ.f<-factor(pdata$occ)
pdata$occ2<-floor(pdata$occ/10)
pdata$occ2.f<-factor(pdata$occ2)
pdata$year.f<-factor(pdata$year)
pdata$marry.yn<-ifelse(pdata$marry!=1, 1, 0)
pdata$before5<-ifelse(pdata$uage<=5,1,0)
#pdata.new<-dummy.data.frame(pdata, sep=".")
pdata$age2<-pdata$age^2
pdata$uprising_e.before5<-pdata$uprising_e*pdata$before5
pdata$uprising_e6.before5<-pdata$uprising_e6*pdata$before5
pdata$feducol<-ifelse(pdata$fedu>=16, 1, 0)
pdata$uprising_e6.before5.fedu<-pdata$uprising_e6*pdata$before5*pdata$fedu
pdata$uprising_e6.before5.feducol<-pdata$uprising_e6*pdata$before5*pdata$feducol
pdata$uprising_e6.feducol<-pdata$uprising_e6*pdata$feducol
pdata$uprising_e6.uage<-pdata$uprising_e6*pdata$uage
pdata$uprising_e6.uage2<-pdata$uprising_e6.uage^2
pdata$uprising_e6.fedu<-pdata$uprising_e6*pdata$fedu

source('d://klips2017/script/4_desc.R')

library(devtools)
install.packages("rlang")
install.packages("bindrcpp")
install.packages("glue")
install.packages("pkgconfig")
install.packages("tibble")
install.packages("mnormt")
install_github("dgrtwo/broom")
library(broom)

#a<-subset(pdata,is.na(lhwage)==F & ((byear>=1982 | byear<=1996) | uprising_e==1) & elechk==1)
a<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) | (byear>=1982 & byear<=1996) | uprising_e==1) & elechk==1)
attach(a)

age.gwangju_e<-age*gwangju_e
age2.gwangju_e<-age2*gwangju_e

#m<-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+edu+occ2.f+bro_num+sis_num+sib_order)
lmfit <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
#lmfit <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu)
#lmfit <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f+age.gwangju_e+age2.gwangju_e)
#m<-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+area.f+fedu+edu+uprising_e6.before5+occ2.f+bro_num+sis_num+sib_order)
#m<-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+area.f+fedu+edu+uprising_e.before5+uprising_e6.before5+uprising_e6.feducol+occ2.f)
#m<-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+area.f+fedu+edu+uprising_e.before5+uprising_e6.before5+occ2.f+bro_num+sis_num+sib_order)
#m<-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+area.f+fedu+edu+uprising_e6.uage+uprising_e6.feducol+occ2.f)
summary(lmfit)
lmfit$vcov <- cluster.vcov(lmfit, ~pid)
lmfith<-coeftest(lmfit, lmfit$vcov)
#tidy_lmfith <- tidy(lmfith)
write.csv(lmfith, "d://klips2017/results/lmfith.csv")

detach(a)
# 
# b<-subset(pdata,is.na(lhwage)==F & ((byear>=1982 & byear<=1996) | (uprising_e==1  & before5==1)) & elechk==1)
# b<-subset(pdata,is.na(lhwage)==F & ((byear>=1982) | (uprising_e==1  & before5==1)) & elechk==1)
b<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | (uprising_e==1  & before5==1)) & elechk==1)
attach(b)
age.gwangju_e<-age*gwangju_e
age2.gwangju_e<-age2*gwangju_e

lmfit2 <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
# lmfit2 <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu)

summary(lmfit2)
lmfit2$vcov <- cluster.vcov(lmfit2, ~pid)
lmfit2h<-coeftest(lmfit2, lmfit2$vcov)
write.csv(lmfit2h, "d://klips2017/results/lmfit2h.csv")
detach(b)

# c<-subset(pdata,is.na(lhwage)==F & ((byear>=1982 & byear<=1996) | (uprising_e==1 & before5==0)) & elechk==1 )
# c<-subset(pdata,is.na(lhwage)==F & ((byear>=1982) | (uprising_e==1 & before5==0)) & elechk==1 )
c<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | (uprising_e==1 & before5==0)) & elechk==1 )
attach(c)
lmfit3 <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
lmfit3 <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu)
summary(lmfit3)
lmfit3$vcov <- cluster.vcov(lmfit3, ~pid)
lmfit3h<-coeftest(lmfit3, lmfit3$vcov)
write.csv(lmfit3h, "d://klips2017/results/lmfit3h.csv")
detach(c)

bdata$female<-ifelse(bdata$sex==2, 1, 0)
bdata$before5<-ifelse(bdata$uage<=5,1,0)
bdata$colyn<-ifelse(bdata$edu>=16, 1, 0)
bdata$feducol<-ifelse(bdata$fedu>=16, 1, 0)

d<-subset(bdata,is.na(edu)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | uprising_e==1) & elechk==1)
attach(d)
lme1<-lm(colyn~uprising_e+gwangju_e+uprising_e6+female+fedu+feducol)
summary(lme1)
# lme1$vcov <- cluster.vcov(lme1, ~pid)
# lme1h<-coeftest(lme1, lme1$vcov)
# write.csv(lme1h, "d://klips2017/results/lme1h.csv")
tidy_lme1 <- tidy(lme1)
write.csv(tidy_lme1, "d://klips2017/results/tidy_lme1.csv")
detach(d)

e<-subset(bdata,is.na(edu)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | (uprising_e==1  & before5==1)) & elechk==1 )
attach(e)
lme2<-lm(colyn~uprising_e+gwangju_e+uprising_e6+female+fedu+feducol)
summary(lme2)
# lme2$vcov <- cluster.vcov(lme2, ~pid)
# lme2h<-coeftest(lme2, lme2$vcov)
# write.csv(lme2h, "d://klips2017/results/lme2h.csv")
tidy_lme2 <- tidy(lme2)
write.csv(tidy_lme2, "d://klips2017/results/tidy_lme2.csv")
detach(e)

f<-subset(bdata,is.na(edu)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | (uprising_e==1 & before5==0)) & elechk==1  )
attach(f)
lme3<-lm(colyn~uprising_e+gwangju_e+uprising_e6+female+fedu+feducol)
summary(lme3)
# lme3$vcov <- cluster.vcov(lme3, ~pid)
# lme3h<-coeftest(lme3, lme3$vcov)
# write.csv(lme3h, "d://klips2017/results/lme3h.csv")
tidy_lme3 <- tidy(lme3)
write.csv(tidy_lme3, "d://klips2017/results/tidy_lme3.csv")
detach(f)

g<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) | (byear>=1982 & byear<=1996) | uprising_e==1) & elechk2==1)
attach(g)
lmfit8 <-lm(lhwage~uprising_e+e13+uprising_e13+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
summary(lmfit8)
lmfit8$vcov <- cluster.vcov(lmfit8, ~pid)
lmfit8h<-coeftest(lmfit8, lmfit8$vcov)
write.csv(lmfit8h, "d://klips2017/results/lmfit8h.csv")
detach(g)

h<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | (uprising_e==1  & before5==1)) & elechk2==1)
attach(h)
lmfit5 <-lm(lhwage~uprising_e+e13+uprising_e13+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
summary(lmfit5)
lmfit5$vcov <- cluster.vcov(lmfit5, ~pid)
lmfit5h<-coeftest(lmfit5, lmfit5$vcov)
write.csv(lmfit5h, "d://klips2017/results/lmfit5h.csv")
detach(h)

i<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | (uprising_e==1  & before5==0)) & elechk2==1)
attach(i)
lmfit6 <-lm(lhwage~uprising_e+e13+uprising_e13+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
summary(lmfit6)
lmfit6$vcov <- cluster.vcov(lmfit6, ~pid)
lmfit6h<-coeftest(lmfit6, lmfit6$vcov)
write.csv(lmfit6h, "d://klips2017/results/lmfit6h.csv")
detach(i)

j<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) | (byear>=1982 & byear<=1996) | uprising_e==1) & elechk==1 & (breg==6 | breg==12 | breg==13))
attach(j)
lmfit7 <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
summary(lmfit7)
lmfit7$vcov <- cluster.vcov(lmfit7, ~pid)
lmfit7h<-coeftest(lmfit7, lmfit7$vcov)
write.csv(lmfit7h, "d://klips2017/results/lmfit7h.csv")
detach(j)

k<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | (uprising_e==1  & before5==1)) & elechk==1 & (breg==6 | breg==12 | breg==13))
attach(k)
lmfit8 <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
summary(lmfit8)
lmfit8$vcov <- cluster.vcov(lmfit8, ~pid)
lmfit8h<-coeftest(lmfit8, lmfit8$vcov)
write.csv(lmfit8h, "d://klips2017/results/lmfit8h.csv")
detach(k)

l<-subset(pdata,is.na(lhwage)==F & ((byear>=1946 & byear<=1960) |  (byear>=1982 & byear<=1996) | (uprising_e==1  & before5==0)) & elechk==1 & (breg==6 | breg==12 | breg==13))
attach(l)
lmfit9 <-lm(lhwage~uprising_e+gwangju_e+uprising_e6+age+age2+year.f+female+marry.yn+fedu+feducol+edu+occ2.f)
summary(lmfit9)
lmfit9$vcov <- cluster.vcov(lmfit9, ~pid)
lmfit9h<-coeftest(lmfit9, lmfit9$vcov)
write.csv(lmfit9h, "d://klips2017/results/lmfit9h.csv")
detach(l)
